Update main.c
Update of the loop while in main.
Update of Screen functions.

Code compiling without mistakes or warnings.

+++ Core/Inc/user_def_sensors.h
@@ -45,6 +45,6 @@ uint8_t determineDirection(uint16_t);
 
 //SENSOR rainfall
 void SENSOR_Rain_Read_Data(void);
-float calculateRainfall(uint16_t);
+float calculateRainfall(uint32_t);
 
 #endif /* INC_USER_DEF_SENSORS_H_ */
